pyproject.toml
[project]
name = "valjean-agent"
version = "0.1.0"
description = "Valjean voice agent worker"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[openai,deepgram,silero,cartesia,noise-cancellation]>=1.0.0",
    "livekit-plugins-turn-detector>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
]
